export compress function
