+++ packages/client/src/components/avatar-panel.tsx
@@ -2,6 +2,7 @@ import { Button } from '@/components/ui/button';
 import type { Agent } from '@elizaos/core';
 import { Image as ImageIcon, Upload, X } from 'lucide-react';
 import { useEffect, useRef, useState } from 'react';
+import { compressImage } from '@/lib/utils';
 
 interface AvatarPanelProps {
   characterValue: Agent;
@@ -12,43 +13,15 @@ export default function AvatarPanel({ characterValue, setCharacterValue }: Avata
   const [avatar, setAvatar] = useState<string | null>(characterValue?.settings?.avatar || null);
   const fileInputRef = useRef<HTMLInputElement>(null);
 
-  const handleFileUpload = (event: React.ChangeEvent<HTMLInputElement>) => {
+  const handleFileUpload = async (event: React.ChangeEvent<HTMLInputElement>) => {
     const file = event.target.files?.[0];
     if (file) {
-      const reader = new FileReader();
-      reader.onload = (e) => {
-        if (e.target?.result) {
-          const img = new Image();
-          img.src = e.target.result as string;
-          img.onload = () => {
-            const canvas = document.createElement('canvas');
-            const maxSize = 300; // Resize to max 300px width/height
-            let width = img.width;
-            let height = img.height;
-
-            if (width > height) {
-              if (width > maxSize) {
-                height *= maxSize / width;
-                width = maxSize;
-              }
-            } else {
-              if (height > maxSize) {
-                width *= maxSize / height;
-                height = maxSize;
-              }
-            }
-
-            canvas.width = width;
-            canvas.height = height;
-            const ctx = canvas.getContext('2d');
-            ctx?.drawImage(img, 0, 0, width, height);
-            const resizedBase64 = canvas.toDataURL('image/jpeg', 0.8); // Reduce quality to 80%
-
-            setAvatar(resizedBase64);
-          };
-        }
-      };
-      reader.readAsDataURL(file);
+      try {
+        const compressedImage = await compressImage(file);
+        setAvatar(compressedImage);
+      } catch (error) {
+        console.error('Error compressing image:', error);
+      }
     }
   };
